requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.0
djangorestframework==3.14.0
mysql==0.0.3
mysqlclient==2.2.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
